feat(decoder): emit speedMs at native 0.5 m/s quantum
V2 decoder rounded byte[7] (int8 x 0.5 m/s) to Int before populating
Vehicle.speedMs. Carry Float through for future trajectory-derived
features.

Safety gates fire byte-for-byte the same; pinned by a -5.5f-vs-6f
regression test.

STATIONARY_SPEED_MS and MOVING_SPEED_MS now correctly reject raw byte
-3 (real 1.5 m/s closing). speedKmh truncates a more-precise Float,
so overlay colour bands switch one quantum earlier in half of cases.
Capture log closing_mps field is now Float.

+++ app/src/main/java/es/jjrh/bikeradar/ClosePassDetector.kt
@@ -42,8 +42,9 @@ class ClosePassDetector {
         /** Minimum closing speed (m/s) for the detector to arm.
          *  Filters out lane-matched cruising and filtering — if the
          *  vehicle isn't genuinely overtaking, it's not a close pass
-         *  event we care to log. */
-        val closingSpeedFloorMs: Int = 6,
+         *  event we care to log. Float at the radar's 0.5 m/s native
+         *  quantum (raw byte * 0.5). */
+        val closingSpeedFloorMs: Float = 6f,
         /** Arm when |rangeX| drops under this. Tiered by rider speed
          *  at ARM time in [decide]: riderSpeed <= 30 km/h uses
          *  [armRangeXUrbanM]; above uses [armRangeXRuralM]. */
